requirements.txt
flask==3.0.3
redis==5.0.6
python-telegram-bot==21.0.1
python-dotenv==1.0.1

